requirements.txt
Django==4.2.13
djangorestframework==3.15.1
drf-spectacular==0.27.2
setuptools>=65.5.1
gunicorn
